Feature/minh.hoang/sence1 1

+++ 05-ScenceManager/Camera.cpp
@@ -53,26 +53,26 @@ void CCamera::Update(CMario* player) {
 	{
 		//vi position cua Camera::GetInstance() ma chinh giua Camera::GetInstance()
 		//luc nay o vi tri goc ben trai cua the gioi thuc
-		SetPosition(GetWidth() / 2, GetPosition().y);
+		SetPosition(GetWidth() / 2.0f, GetPosition().y);
 	}
 
 	if (GetBound().right > mapWidth)
 	{
 		//luc nay cham goc ben phai cua the gioi thuc
-		SetPosition(mapWidth - GetWidth() / 2, GetPosition().y);
+		SetPosition(mapWidth - GetWidth() / 2.0f, GetPosition().y);
 	}
 
 	if (GetBound().top < 0)
 	{
 		//luc nay cham goc tren the gioi thuc
-		SetPosition(GetPosition().x, GetHeight() / 2);
+		SetPosition(GetPosition().x, GetHeight() / 2.0f);
 	}
 
 	if (GetBound().bottom > 300)
 		/*if (GetBound().bottom > mapHeight / 2)*/
 	{
 		//luc nay cham day cua the gioi thuc
-		SetPosition(GetPosition().x, mapHeight - GetHeight() / 2);
+		SetPosition(GetPosition().x, mapHeight - GetHeight() / 2.0f);
 	}
 
 }
@@ -87,10 +87,10 @@ RECT CCamera::GetBound()
 {
 	RECT bound;
 
-	bound.left = position.x - width / 2;
-	bound.right = bound.left + width;
-	bound.top = position.y - height / 2;
-	bound.bottom = bound.top + height;
+	bound.left = (double)(position.x - width / 2);
+	bound.right = (double)(bound.left + width);
+	bound.top = (double)(position.y - height / 2);
+	bound.bottom = (double)(bound.top + height);
 
 	return bound;
 }
